Refactor mnist examples and e2e test scripts
- Unify many copies of mnist training scripts into a single one.
  - Backends: Gloo, NCCL, MPI.
  - Bases of the official pytorch/examples/mnist example.
  - Uses TensorboardX for summary writing.
  - Uses a backend argument to select backend (if applicable).
- Update all examples and dockerfiles to Pytorch 1.0.
- Remove examples using deprecated TCP backend.
- Remove old v1alpha1 examples.
