--- examples/ddp/mnist/cpu/v1beta1/job_mnist_ddp_cpu.yaml
@@ -1,22 +1,0 @@
-apiVersion: "kubeflow.org/v1beta1"
-kind: "PyTorchJob"
-metadata:
-  name: "pytorch-mnist-ddp-cpu"
-spec:
-  pytorchReplicaSpecs:
-    Master:
-      replicas: 1
-      restartPolicy: OnFailure
-      template:
-        spec:
-          containers:
-            - name: pytorch
-              image: docker.io/andreyvelichkevich/pytorch-mnist-ddp-cpu
-    Worker:
-      replicas: 3
-      restartPolicy: OnFailure
-      template:
-        spec:
-          containers:
-            - name: pytorch
-              image: docker.io/andreyvelichkevich/pytorch-mnist-ddp-cpu
