--- examples/ddp/mnist/gpu/v1beta1/job_mnist_ddp_gpu.yaml
@@ -1,29 +1,0 @@
-apiVersion: "kubeflow.org/v1beta1"
-kind: "PyTorchJob"
-metadata:
-  name: "pytorch-mnist-ddp-gpu"
-spec:
-  pytorchReplicaSpecs:
-    Master:
-      replicas: 1
-      restartPolicy: OnFailure
-      template:
-        spec:
-          containers:
-            - name: pytorch
-              image: docker.io/deepermind/pytorch-mnist-ddp-gpu
-              resources:
-                limits:
-                  nvidia.com/gpu: 1
-    Worker:
-      replicas: 3
-      restartPolicy: OnFailure
-      template:
-        spec:
-          containers:
-            - name: pytorch
-              image: docker.io/deepermind/pytorch-mnist-ddp-gpu
-              resources: 
-                limits:
-                  nvidia.com/gpu: 1
-
