+++ examples/mnist/Dockerfile-mpi
@@ -1,4 +1,4 @@
-FROM nvidia/cuda:latest
+FROM nvidia/cuda:9.0-cudnn7-devel-ubuntu16.04
 ARG PYTHON_VERSION=3.6
 
 RUN apt-get update && apt-get install -y --no-install-recommends \
@@ -30,7 +30,6 @@ RUN curl -o ~/miniconda.sh -O  https://repo.continuum.io/miniconda/Miniconda3-la
      rm ~/miniconda.sh && \
      /opt/conda/bin/conda update conda && \
      /opt/conda/bin/conda install -y python=$PYTHON_VERSION numpy pyyaml scipy ipython mkl mkl-include cython typing && \
-#     /opt/conda/bin/conda install -c conda-forge openmpi && \
      /opt/conda/bin/conda clean -ya
 ENV PATH /opt/conda/bin:$PATH
 # This must be done before pip so that requirements.txt is available
@@ -45,9 +44,9 @@ RUN TORCH_CUDA_ARCH_LIST="3.5 5.2 6.0 6.1 7.0+PTX" TORCH_NVCC_FLAGS="-Xfatbin -c
 
 RUN /opt/conda/bin/conda config --set ssl_verify False
 RUN pip install --upgrade pip --trusted-host pypi.org --trusted-host files.pythonhosted.org
-RUN pip install --trusted-host pypi.org --trusted-host files.pythonhosted.org torchvision
-WORKDIR /workspace
-RUN chmod -R a+w /workspace
-ADD . /opt/pytorch_dist_mnist
+RUN pip install --trusted-host pypi.org --trusted-host files.pythonhosted.org torchvision tensorboardX==1.6.0
 
-ENTRYPOINT ["mpirun", "-n", "1", "--allow-run-as-root", "python", "-u", "/opt/pytorch_dist_mnist/mpi_mnist_gpu.py"]
+WORKDIR /var
+ADD mnist.py /var
+
+ENTRYPOINT ["mpirun", "-n", "1", "--allow-run-as-root", "python", "/var/mnist.py"]
