+++ examples/mnist/Makefile
