+++ examples/mnist/v1beta1/pytorch_job_mnist_mpi.yaml
@@ -1,0 +1,32 @@
+apiVersion: "kubeflow.org/v1beta1"
+kind: "PyTorchJob"
+metadata:
+  name: "pytorch-dist-mnist-mpi"
+spec:
+  pytorchReplicaSpecs:
+    Master:
+      replicas: 1
+      restartPolicy: OnFailure
+      template:
+        spec:
+          containers:
+            - name: pytorch
+              image: gcr.io/<your_project>/pytorch_dist_mnist:mpi
+              args: ["--backend", "mpi"]
+              # Comment out the below resources to use the CPU.
+              resources: 
+                limits:
+                  nvidia.com/gpu: 1
+    Worker:
+      replicas: 1
+      restartPolicy: OnFailure 
+      template:
+        spec:
+          containers: 
+            - name: pytorch
+              image: gcr.io/<your_project>/pytorch_dist_mnist:mpi
+              args: ["--backend", "mpi"]
+              # Comment out the below resources to use the CPU.
+              resources: 
+                limits:
+                  nvidia.com/gpu: 1
