--- examples/mpi-dist/mnist/cpu/v1alpha2/mpi_mnist_job_cpu.yaml
@@ -1,22 +1,0 @@
-apiVersion: "kubeflow.org/v1alpha2"
-kind: "PyTorchJob"
-metadata:
-  name: "pytorch-mpi-mnist-cpu"
-spec:
-  pytorchReplicaSpecs:
-    Master:
-      replicas: 1
-      restartPolicy: OnFailure
-      template:
-        spec:
-          containers:
-            - name: pytorch
-              image: gcr.io/kubeflow-ci/pytorch-mpi-mnist-cpu:1.0
-    Worker:
-      replicas: 3
-      restartPolicy: OnFailure
-      template:
-        spec:
-          containers:
-            - name: pytorch
-              image: gcr.io/kubeflow-ci/pytorch-mpi-mnist-cpu:1.0
