--- examples/tcp-dist/cifar10/Dockerfile
@@ -1,4 +1,0 @@
-FROM pytorch/pytorch:0.4_cuda9_cudnn7
-
-ADD . /opt/pytorch_dist_cifar
-ENTRYPOINT ["python", "/opt/pytorch_dist_cifar/dist_cifar.py"]
