--- examples/tcp-dist/cifar10/v1alpha1/pytorch_job_cifar.yaml
@@ -1,26 +1,0 @@
-apiVersion: "kubeflow.org/v1alpha1"
-kind: "PyTorchJob"
-metadata:
-  name: "pytorch-dist-cifar"
-spec:
-  backend: "tcp"
-  masterPort: "23456"
-  replicaSpecs:
-    - replicas: 1
-      replicaType: MASTER
-      template:
-        spec:
-          containers:
-          - image: gcr.io/kubeflow-ci/pytorch-dist-cifar-test:1.0
-            imagePullPolicy: Always
-            name: pytorch
-          restartPolicy: OnFailure
-    - replicas: 3
-      replicaType: WORKER
-      template:
-        spec:
-          containers:
-          - image: gcr.io/kubeflow-ci/pytorch-dist-cifar-test:1.0
-            imagePullPolicy: Always
-            name: pytorch
-          restartPolicy: OnFailure
