--- examples/tcp-dist/cifar10/v1alpha2/pytorch_job_cifar.yaml
@@ -1,22 +1,0 @@
-apiVersion: "kubeflow.org/v1alpha2"
-kind: "PyTorchJob"
-metadata:
-  name: "pytorch-dist-cifar"
-spec:
-  pytorchReplicaSpecs:
-    Master:
-      replicas: 1
-      restartPolicy: OnFailure
-      template:
-        spec:
-          containers:
-            - name: pytorch
-              image: gcr.io/kubeflow-ci/pytorch-dist-cifar-test:1.0
-    Worker:
-      replicas: 3
-      restartPolicy: OnFailure
-      template:
-        spec:
-          containers:
-            - name: pytorch
-              image: gcr.io/kubeflow-ci/pytorch-dist-cifar-test:1.0
