--- examples/tcp-dist/mnist/Dockerfile
@@ -1,4 +1,0 @@
-FROM pytorch/pytorch:0.4_cuda9_cudnn7
-
-ADD . /opt/pytorch_dist_mnist
-ENTRYPOINT ["python", "/opt/pytorch_dist_mnist/dist_mnist.py"]
