--- examples/tcp-dist/mnist/v1beta1/pytorch_job_mnist.yaml
@@ -1,22 +1,0 @@
-apiVersion: "kubeflow.org/v1beta1"
-kind: "PyTorchJob"
-metadata:
-  name: "pytorch-tcp-dist-mnist"
-spec:
-  pytorchReplicaSpecs:
-    Master:
-      replicas: 1
-      restartPolicy: OnFailure
-      template:
-        spec:
-          containers:
-            - name: pytorch
-              image: gcr.io/kubeflow-ci/pytorch-dist-mnist_test:1.0
-    Worker:
-      replicas: 3
-      restartPolicy: OnFailure
-      template:
-        spec:
-          containers:
-            - name: pytorch
-              image: gcr.io/kubeflow-ci/pytorch-dist-mnist_test:1.0
